+++ Assets/_Scripts/Systems/Events/Tracker/EventsTracker.cs
@@ -1,56 +1,88 @@
-using System;
 using System.Collections.Generic;
-using System.Text;
+using Metroidvania.Events.Tracker.Handles;
 using UnityEngine;
+using System;
 
 namespace Metroidvania.Events.Tracker
 {
     // TODO: Add editor UI and add an text file in runtime to handle events tacking
     public class EventsTracker : ScriptableSingleton<EventsTracker>
     {
-        public struct EventTrack
+        [System.Flags]
+        public enum TracksHandler
         {
-            public EventChannelBase channel;
+            File = 1 << 0,
+        }
 
-            public double invokeTime;
+        public class EventTrack : IDisposable
+        {
+            public readonly EventsTracker tracker;
+            public readonly EventChannelBase channel;
+
+            public System.Action<EventTrack> trackReleased;
+
+            public double invokeTime, releaseTime;
 
             public object[] invokeParams;
-        }
 
-        public System.Action<EventTrack> EventRaised;
-        public bool trackingEnabled;
-
-        private void OnEnable()
-        {
-            EventRaised += OnEventRaise;
-        }
-
-        private void OnDisable()
-        {
-            EventRaised -= OnEventRaise;
-        }
-
-        private void OnEventRaise(EventTrack track)
-        {
-            StringBuilder argsSb = new StringBuilder();
-            for (int i = 0; i < track.invokeParams.Length; i++)
+            public EventTrack(EventsTracker tracker, EventChannelBase channel, System.Action<EventTrack> trackReleased, object[] invokeParams)
             {
-                object invokeParam = track.invokeParams[i];
-                argsSb.AppendFormat("\n  - Arg {0}: {1} ({2})", i, invokeParam, invokeParam.GetType());
+                this.tracker = tracker;
+                this.channel = channel;
+                this.invokeParams = invokeParams;
+                this.trackReleased = trackReleased;
+                this.invokeTime = Time.realtimeSinceStartupAsDouble;
+                this.releaseTime = -1;
             }
 
-            Debug.Log($"The channel {track.channel.name} ({track.channel.GetType().Name}) has raised an event at ({track.invokeTime}). {argsSb}");
+            public void Dispose()
+            {
+                releaseTime = Time.realtimeSinceStartupAsDouble;
+                trackReleased?.Invoke(this);
+            }
         }
 
-        private List<EventChannelBase> _trackingChannels = new List<EventChannelBase>();
+        public static System.Action<EventTrack> OnBeginEventTrack;
+        public static System.Action<EventTrack> OnEndEventTrack;
 
-        public void TrackEvent(EventChannelBase channel, object[] args)
+        [SerializeField] private bool m_trackingEnabled;
+        [SerializeField] private TracksHandler m_enabledHandlers;
+
+        public bool trackingEnabled => m_trackingEnabled;
+        public TracksHandler enabledHandlers => m_enabledHandlers;
+
+        private List<IEventTrackerHandler> _handlers;
+
+        public EventTrack BeginEventTrack(EventChannelBase channel, object[] args)
         {
-            if (!trackingEnabled) return;
+            var eventTrack = new EventTrack(this, channel, EventTrackReleased, args);
 
-            var stackTrace = new EventTrack() { channel = channel, invokeTime = Time.realtimeSinceStartupAsDouble, invokeParams = args };
+            if (_handlers != null)
+            {
+                foreach (var handle in _handlers)
+                    handle.BeginTrack(eventTrack);
+            }
 
-            EventRaised?.Invoke(stackTrace);
+            OnBeginEventTrack?.Invoke(eventTrack);
+            return eventTrack;
+        }
+
+        private void EventTrackReleased(EventTrack track)
+        {
+            if (_handlers != null)
+            {
+                foreach (var handle in _handlers)
+                    handle.EndTrack(track);
+            }
+
+            OnEndEventTrack?.Invoke(track);
+        }
+
+        public void AddHandler(IEventTrackerHandler handler)
+        {
+            if (_handlers == null)
+                _handlers = new List<IEventTrackerHandler>();
+            _handlers.Add(handler);
         }
     }
 }