+++ Assets/_Scripts/Systems/Events/Tracker/IEventTrackerHandler.cs
@@ -1,0 +1,8 @@
+namespace Metroidvania.Events.Tracker.Handles
+{
+    public interface IEventTrackerHandler
+    {
+        void BeginTrack(EventsTracker.EventTrack track);
+        void EndTrack(EventsTracker.EventTrack track);
+    }
+}